Add status indicators to main page
In order to help debugging the PWA, specially the service worker part
which is a bit tricky and complex, some status indicators are added to
the main page.

For now, only indicators about the service worker are present, but in
the future the same framework may be used for other indicators. Also the
entire indicator area can be hidden in production, if needed, it's all
in a single <div>.

The changes in the service worker code are not exactly needed for this,
but they have been added to this commit because they were needed for
testing the indicator system, so it makes sense to commit them now.