Add namespace for project and adapt the code to it
A namespace is created for the application. It will be used for the main
application code, too, to avoid pollution of global namespace.

Some identifiers were adapted to CamelCase, too. Looks like it is the
standard for JavaScript, so using it for this project makes the code
look more similar to standard and library code.