Fix precaching when serving path is not root
Turns out 'cache.addAll()' can fail if the PWA is served served from a
path (and thus, scope) other than root if the paths for the assets are
specified using absolute paths. This happens because absolute paths are
resolved from the root of the URL, not from the scope.

In order to avoid this, all precached assets are now specified using
relative paths instead of absolute ones.

The only problem with this approach is how to specify a relative route
for the root of the PWA, the index HTML. For now, the PWA scope is used
for that path, although a simpler alternative exists.