Improve service worker speed
For the moment, fetches handled by the service worker will be filtered
based upon the method (only GET requests will be processed) and origin
(only same-origin requests will be processed). Any other request will be
handled by the browser, not the service worker.

In reality the speed improvement is marginal but the responsiveness is
not, and in any case (and for the time being) it's the way to go.