Modify main page look
In order to implement the definitive page look, the first step is making
it responsive, fail gracefully when JavaScript is not enabled, and adapt
the contents to make room for the elements that will be added later.

So, the status box is only shown when JavaScript is enabled and working,
the HTML has been cleaned up and reorganized, with a new header, and the
styles have been modified accordingly.