Add a new "console" HTML element
This element will serve as the communication channel from the PWA to end
users, to report progress, print results, etc.

In the future, and depending on the kind of reporting needed, this may
get simpler or more powerful. For now the idea is to replace the status
component and to use the "console" to communicate with the end user.