Create new page layout
From now on a new page layout is used. The old status area on the main
page has been replaced by the new "console" element and all temporary
debugging information goes there, now.

Also the size of the new page has changed and a new dynamic layout is
used, where the "console" fills all available vertical space if needed
in a flexible manner.

The main JavaScript code now uses the "console" element, too, and the
old function to show debugging status has been deprecated. The display
of debugging information is by no means definitive, as in production the
messages output to the end user should be kept to a minimum, so maybe
making the entire "console" element unnecessary.