Replace version number strategy
After much thinking, the version number is now stored within the service
worker, not the main JavasScript code. The main reason is consistency,
because with a 'cache-only' caching strategy, having only static core
assets, a version of the PWA consists on what exists on a version of the
core assets cache, and that is defined by the service worker version.

On the other hand, having the version number inside the service worker
makes sure the cache is properly updated when increasing the version of
the PWA, since the service worker will be modified, reducing the risk of
having a half-baked version of the PWA on production because the service
worker was not modified by mistake.

So, a new version number handling was implemented, on a fake URL called
"version", on the root of the scope. Requesting that URL returns the
current version number within the service worker, as plain text.