Create simple feature detection
For now, only the File API is feature-detected using this new system of
hiding elements on the main page until they are safe to show, but it can
be used for other features.

Not all features prevent the PWA from working, so not all of them will
be handled like this.