Replace file loader UI element
Previously, a well-known approach was used: using a form, containing a
hidden input element with a label attached so it could be styled with
CSS, tab-navigated using keyboard, etc.

Unfortunately that was unnecessary complex and also the layout was not
optimal, not symmetric and... well... ugly.

So the entire element has been replaced with a simpler and maybe more
straightforward solution, using div elements, symmetric buttons, etc.

The basic logic to handle this new UI combination has been added to the
main JavaScript code, too.