New layout for communicating with end user
The previous "console" element was not enough to contain debugging, log
and status messages, and also debugging messages should appear grouped
and isolated so they can be hidden in production.

Also, increased flexibility is needed for printing messages on the main
page, while at the same time the printing function itself has to be made
faster and simpler.

For those reasons a new layout has been designed.

Now, there are two separate "consoles". The first one will be used for
communication with the end user, for status, progress and error messages
needed during normal functioning. The second one will be the logging and
debugging console, and will be functional only during development.

Last but not least, a generic method to stop execution in case of fatal
errors (most of the time, missing features needed for operation) is now
provided.

As a result, CSS is a bit simpler, too.