Improve missing feature detection
By separating the error logging mechanism from the termination logic,
it is more flexible, a bit simpler and feature detection is easier and
more compact.

Also, it makes much more sense: now there are three separate logging
functions, one for debugging, one for normal messages and another for
error messages, which are more or less symmetrical. And separated from
those, there is an execution termination function.