Fix HTML element width
Previous CSS styling assigned a 100vw witdh to the HTML element, to fill
the entire viewport width. Unfortunately this means that if the vertical
scrollbar is visible, then an horizontal scrollbar appears, no matter if
it is needed or not, because the width of the vertical scrollbar is not
substracted from the viewport width.

Summing up, the HTML element was wider than the viewport if the vertical
scrollbar appeared.