Change how page elements are hidden
Until now, the 'hidden' property of 'HTMLElement' was used to hide some
page elements until they were needed. Unfortunately, this property is a
bit... weak. Any change on the 'display' property will override 'hidden'
so it's useless for the intended purpose.

Using a class to set 'display' to 'none', and removing that class using
JavaScript is a much better solution because that way it's not essential
to remember the particular value of 'display' for any element that must
be hidden at some point.

See https://css-tricks.com/the-hidden-attribute-is-visibly-weak/ for
additional details.