Simplify console and logging. Heavily...
Previously, the so-called 'console' was double and quite a bit complex
because it has to be able to handle complex logging and debugging, just
in case it was needed in the future, specially for the file conversion
user interface, as a means to communicate results to the end user while
at the same time having a logging area for debugging, etc.

That is no longer needed, and in fact a simpler user interface is wanted
in production, so now there's a single 'console', for logging mainly, so
it won't be much used in production except to communicate fatal errors
to the end user.

As a result, the markup, the styles and the code have been simplified,
heavily, and a way of disabling the whole logging facility in production
has been added, too.