Use new way of hidding/showing elements
Previously a class was used to hide and show elements as needed, but the
problem is that it's more verbose than using the 'hidden' HTML attribute
and it doesn't solve the specificity problem, since a class selector is
in the end less specific than an id selector, forcing the use of complex
selectors or '!important'.

So, since the problems and the solution is the same, the 'hidden' HTML
attribute is now used to hide and show elements. Using '!important' in
the CSS special selector for 'hidden' is clearer than using it on class
selectors, and resulting HTML is semantically better and more compact.