Improve layout responsiveness
Before, in order to have a responsive but visually pleasing layout, the
minimum width of the body was made wide enought to fit all content, no
matter how wide it was. It worked.

Unfortunately this creates a huge problem with very long filenames when
the screen width is not that big: most of the user interface, being as
it is, centered, was invisible unless the user scrolled horizontally.

This happened because the very long filename caused that element to be
very wide, and the text never wrapped because the body element had been
instructed to be as wide as needed to fit the content. So, the very wide
element forced the ENTIRE page to be extremely wide and that caused that
ALL centered content needed horizontal scroll to be visible. Not funny.

This has been fixed by forcing all the text on the page to wrap when the
viewport width decreases, allowing all content to be visible because the
page can be now be incredibly narrow. Of course, this UI is not the most
usable in extreme cases (viewport width well below 200px), but that will
not be a common case, and this way the user interface is more flexible.

This may need further refinement in the future.