Fix text wrapping in flexboxed paragraphs
Paragraphs within flex containers did not wrap their text properly due
to the fact that the container (being flexbox) did not impose a minimum
width on the paragraphs.

Before the solution was to use 'overflow: hidden;'. Frankly, no idea why
this worked, but it's not the proper solution. The proper solution is to
force a minimum width on the paragraph as little as possible. So, 0px.