Fix problem with button outline
Apparently, modern Firefox browsers (>88) will render the outline of all
elements using border-radius into account. So, buttons in Firefox would
have an outline with rounded corners and all other browsers will render
the outline with square corners.

The solution is to remove border-radius in the button states which will
show an outline.