Minimize usage of trujamanConsole.log()
This is the first step towards getting rid of trujamanConsole as it is
now, since it's quite complex and it was created (mainly) to inform the
user about progress and other states belonging to conversion jobs.

Since the user interface has changed, trujamanConsole makes no sense now
because it will be only used to communicate fatal errors to the user but
no informational messages.