Simplify service worker registration handler
This is really a two-sided change. On one hand, it is the second step
towards replacing trujamanConsole. On the other hand, installation of a
new service worker is automatic, no matter if it is intrusive, for the
time being. So, handling each and every state in the service worker's
lifecycle is absolutely overkill (again, for the time being).