Simplify install prompt handling
This is, again, a two pronged change. On one hand is the last step for
getting rid of trujamanConsole (well, apart from actually getting rid of
trujamanConsole code itself, of course). On the other hand, since the
people at Mozilla decided that PWA won't be installable anymore on their
Firefox browser, a future strategy about installing the PWA where it is
supported must be designed. So, for now it is better to just prevent the
default installation prompt to appear.