Get rid of trujamanConsole
The trujamanConsole component was obsolete so it has been replaced by a
much simpler mechanism of error reporting, since log messages for users
won't be needed anymore.

That meant not only getting rid of the trujamanConsole code itself, but
also rewriting error reporting and modifying both the HTML and the CSS
so everything uses the new system.