Make version visible only when available
This avoids a very ugly empty space caused by the gap in the page header
flexbox when the service worker is not ready or an error occurs. In that
case only the page logo has to appear, but since the version code would
be empty, the gap of the flexbox appeared. Not anymore.