Simplify (again) error notification
It does not make sense to throw an error to try to stop the JavaScript
engine, since ONLY the thread where that error is thrown will be stopped
while others keep running.

So, the error handling will be redesigned, and the first step is making
the error notification function simpler, not affecting execution flow.