Implement background tasks
Now, all potentially long operations will be handled by a Web Worker,
starting with reading files from disk. In the future, more operations
will be handled by the Web Worker.

For now, all file reading has been rewritten, and the low level handling
has been moved to a separate script for the Web Worker.

This has meant a heavy rewrite of the code base because a lot of code
had to be moved and some other code adapted to the new system, which
involves a communication protocol between the main thread and the Web
Worker code.