Return a single byte after file read
During development, and until further file processing is done, the web
worker will return a single byte as payload after file reading. This is
enough to test the execution flow, coverage, etc. while at the same time
reduces complexity and memory usage, and it's data which is not used now
anyway, so...