Delay file reading operations
During development, and in order to properly test the UI, file reading
operations have to be a tad slower. So, every file reading operation is
delayed for 500 milliseconds.

Reading a single file may generate many reading operations, so the total
delay for reading a file may be well above 500 milliseconds.