Improve hashing of existing jobs in web worker
Instead of using the file name, since there is no path information, now
a mix of file name, size and last modification time is used. This is not
exactly bulletproof but it is a huge improvement and anyway the worker
will not get duplicate files so it is safe enough.