Remove feature detection
In general, feature detection has to be used in order to run some code
if a feature is available and run a different code if it's not. That was
not the reason why feature detection was used in this app.

Feature detection was used for three reasons in this app:
1. For detecting if latest syntax can be used.
2. For detecting if some latest feature is available in the browser.
3. For detecting if some available feature has been disabled.

The first reason is very important, since using latest syntax can give
syntax errors if the browser doesn't support that syntax. Since the app
can't work without that syntax, a better solution is to use it anyway
and trap syntax errors so the app doesn't run if they happen.

The second reason is very important too, but the features needed by the
app to work are widely supported and if the browser doesn't support them
an exception will be thrown and properly handled anyway, excepto maybe
in corner cases that will be dealt with in the future, as needed.

So, the main reason why feature detection was used is the third one. The
user can optionally disable some existing and supported features in the
browser. Mainly, cookies. If cookies are disabled by the user, then the
service worker can't be registered, and the app won't work.

Keeping feature detection just for this reason was not good, and in fact
for things like cookies that detection can be done at the proper place
and time, allowing for a better reporting for the end user.

Last but not least, feature detection was breaking encapsulation because
error reporting had to be kept separated from the main user interface
code, even though it should be within the user interface code.