Create UI class, first step towards Passive View
The app is complex enough now to need another design pattern. The chosen
one is Passive View... ish. That is, there's going to be a view (to wit
the UI class) and a presenter, but the model itself is so simple that no
class will be created for it yet.

The UI class encapsulates the communication with the HTML code, for now.

The first step is creating the UI class and adding the first method, one
for showing the version code on the web page.