Add function showError() to UI class
This way, error reporting is now encapsulated in the UI class instead of
being just a helper function.

Since there's no presenter (yet), the UI object has to be passed around
so other objects can use showError(). This is not clean at all, but will
be solved when a presenter is implemented.