Refactor more code into UI class
This time is the file picker related code. Since there is no presenter
available yet, some tweaks had to be performed to keep the code working,
namely in the createJobs() function, which has to be stored inside the
window object to be accessible from the UI class.

That will be fixed soon.