Add Observer pattern to View component
Until now, for the first implementation of the Passive View pattern, the
View component had knowledge of the internals of the Presenter. In fact,
the View KNEW about a Presenter.

From now on, the View just notifies its Observer about user interactions
and lets the Observer handle the events without further knowledge.

In order to communicate the View with its Observer, a quite crude event
handling system has been implemented.