Replace web worker communication protocol
The next step in implementing the Passive View pattern is incorporating
the web worker interface into the Presenter component. That interface
was a promisified version of the postMessage() communication protocol,
which was perfect for mimicking a synchronous call-reply scheme, where
the calls always originated in the Presenter as messages sent to the web
worker, and then the web worker returned the replies asynchronously as
messages sent to the Presenter.

Unfortunately, that didn't allow for the situation where the web worker
had to send messages to the Presenter that weren't replies. For example,
a progress indicator where the web worker sends messages periodically to
the Presenter to indicate the progress of a long-running operation.

Since those messages/events were not being sent as replies to messages
sent from the Presenter to the web worker, the promisified interface
doesn't work, as Promises can only be settled once.

For that reason, a new protocol has been created, which is absolutely
asynchronous and doesn't really have state. The Presenter will send the
messages to the web worker without waiting for a reply. All replies will
be messages sent asynchronously by the web worker to the Presenter.

Both the web worker and the Presenter will maintain respective states to
perform their tasks, but it won't be shared, it will be independent.