Improve error showing code in user interface
Before, when the View component showed an error, a crude method was used
to stop the user interface so further user interaction was impossible.

The method used was to hide all elements, but running jobs kept running
and it was not entirely impossible to cause more errors, overwriting the
one currently shown.

Now, the jobs are properly cancelled and dismissed and after that the
file picker is hidden, so no further interaction is possible. Also, it's
impossible that subsequent errors overwrite the first error shown.