Improve job creation protocol
Before, the job id was passed to the web worker when it created a job,
and that id was computed in the Presenter component. This doesn't make
much sense, as the web worker is the one in charge of actually creating
jobs. The problem was that job duplication was checked it the Presenter
so the job ids had to be known there BEFORE job creation.

Now, job duplication is checked in the web worker and the job ids are
computed there, too.