Add type="button" to HTML buttons
By default, HTML buttons have type 'submit', not 'button', and this may
cause problems in the future. Not now, because now the buttons are all
handled by JavaScript, but maybe in the future there could be problems
with some button acting as type="submit", causing surprising keyboard
behavior in forms.