Prevent content shift due to vertical scroll bar
When the vertical scroll bar appears, the entire content is shifted left
by the amount corresponding to the width of that scroll bar.

This is fixed by assigning the HTML element a width corresponding to the
entire width of the viewport ignoring the scroll bar width, and elements
are shrunk a bit to fit even when the scroll bar appears.