Replace Observer pattern with Publish/Subscribe
The View and the Presenter components were bound by an Observer pattern,
where the Presenter observed the View and in turn invoked View methods.
So, the View had to have a Presenter instance for communication.

That has been replaced by a Publish/Subscribe pattern, so the View is no
longer coupled to the Presenter. The Presenter subscribes to the events
instead and the View does not know anything about how the events it may
emit are handled and by whom.

Of course, this meant removing more logic from the View and that is the
reason why a lot of code was modified.