Fix problem with html element 'height' property
The 'height' property for the html element was set to '100%', and this
means that, with no jobs, the computed height will be that of the entire
viewport, including body padding and margins. Correct.

But when jobs are added and the page scrolls, the html element height is
still the same as before! So, the last job doesn't have any bottom side
margin or padding.

Without a 'height' property, the computed height for the html element is
'auto', and will include any padding and margin even when scrolling.