Add support for reporting more than one error
Before, only one error could be shown on the page. This was intentional
since this reporting was not for end users but for developers, and that
allowed focusing and fixing one error at a time.

Unfortunately, for some errors an error cascade could happen and if so,
only the first error in the cascade would be shown while the rest could
appear as an alert() message or just disappear.

Even though that situation will seldom happen, now all errors happening
during the entire life of the application will appear on the page, no
matter if they will be dealt with one by one.