Improve handling of color schemes in CSS
Before, media queries for both light and dark color schemes were used,
so there was not really a default color scheme. That is, if neither dark
nor light color scheme was set by the user agent or the operating system
then the application will use the color scheme used by the default user
agent stylesheet (usually black and white, no colors).

Now, the default color scheme, if none specified, is light, and it will
be used even if future color schemes are added to the specification. The
exception is dark color scheme, which is covered by a media query, just
like before.

Also, superfluous comments have been removed in that part.