Use <template> instead of hidden <div> for jobs
A hidden <div> element was being used as template for dinamically create
new jobs at runtime. This was achieved by cloning the hidden <div>, then
adding it to a container and finally making it visible.

This was working perfectly, but now a <template> element is used because
it is more future-proof, it is not parsed until used, does not have side
effects until used, etc.

Since the <template> element is not parsed until instantiated, the list
of known formats cannot be populated until that moment, so a very minor
modification has been added to that part of the code, so not it is more
robust and future-proof, too.