Send errors to browser inspector console
Before, errors were only shown in the web page itself. This is clean but
has the nasty issue of potentially hide secondary errors, like the ones
produced in the error-showing code, for example.

By allowing all errors to go to the browser inspector console, there is
at least a last-resort system to access them if necessary, while at the
same time not polluting the user interface (the web page itself).