Rewrite unexpected errors handling
Before, and since these kind of errors could happen well before the page
was fully loaded and working, unexpected errors handling had two parts,
one residing in the HTML file, in the form of a simple embedded script
with a crude message handling mechanism, and the other one when the page
was up and running, more beautiful and well presented.

It was a bit of a mess so it has been rewritten and sanitized. Now all
the code resides in the JavaScript main file and it is more flexible and
with better and more detailed information for the end user.