Sanitize repository
For a start, cSpell ignored words are handled now in a different way, by
adding them as comments with directives for the spell checker (cSpell in
this case), instead of using the IDE local settings file.

Also, the 'zhttpd.cmd' file, containing the command to run a local HTTP
server for development and debugging purposes, has been removed: it has
nothing to do with the repository itself, but rather with a particular
development environment. It is better to remove it.