Fix lack of difference between job objects and IDs
In the code, job DOM objects (belonging to the View) and job IDs (which
are used by the Web Worker) were somewhat intermixed, considering that
BOTH were some form of ID. But DOM objects are not exactly that and they
are the REAL job entity for the user. The IDs needed by the Web Worker
are there just for convenience as DOM objects cannot be passed safely to
a Web Worker.

So some variables have been renamed to make clear if a job DOM object or
a Web Worker job ID is being used.