Use .git/info/exclude rather than .gitignore
The .gitignore file should be used for files that EVERY developer who is
working on the project will want to ignore, such as the files produced
by the building process, but not for local development files. But it was
being used exactly for that, for ignoring the configuration and support
files used by MY local development environment.

That kind of 'ignoring' has to be done using .git/info/exclude, instead.