Add version file
Until now, the version code was hardcoded into the service worker script
because the service worker was conceptually regarded AS the PWA.

While this is not entirely wrong, the approach is not flexible, and the
version code may be needed in other places of the code in the future.

Also, a string-only version code is useful, but since it follows SemVer
it is better if it is structured and stored in an object, which can be
converted into a string when the context requires it.

So, now the version code is in its own file, and this file also provides
a flag to check if debug mode is enabled. That flag is set automatically
for versions with a prerelease identifier.