Delay file reading in debug mode only
Before, file reading operations were unconditionally delayed, since the
PWA is still not mature for production. But this made speed tests quite
cumbersome to perform since the code had to be modified temporarily for
carrying out file reading speed tests.

Also, delaying file reading operations is useful even when the PWA will
be in production, so a way of selectively enabling and disabling those
delays is used now.