Disable auto debug mode for release candidates
For release candidates it is better to switch, at least partially, into
production mode to check if code which only runs in debug mode has some
side effect or is affecting PWA stability.

Of course, this should never happen, but it is better to check for this
before releasing the PWA stable version, in the release-candidate cycle.