Reduce maximum allowed file size to 99 MiB
Actual maximum file size which the app can handle depends, heavily, on
the browser, platform, ECMAScript engine, etc. and can range from 1GiB
to nearly 8GiB. That is excessive for common usage and it is difficult
to know the exact amount in advance.

So, since files handled by the app will seldom be above a few MiB, a new
limit of 99 MiB is set. It is more than enough and it is well known in
advance.