Remove 'version.js' file
The version code has been moved back into the service worker because in
the end it was being used only there, while the rest of imports of the
'version.js' file were for using the 'DEBUG' constant which is no longer
used in the code.

So, and since in fact the service worker can be actually considered as
the app, version-wise, it has been moved back.

For now, the structuring is kept.