Fix race condition in job cancellation
When a job was cancelled and the request was sent to the web worker, as
soon as the request was made by the user, the buttons of the UI element
for the job were set up as if the cancellation had already been done.

Problem is, sometimes the cancellation request was made just a tiny bit
before the file read was complete, and so the file was completely read
by the time the cancellation was actually performed. In short, the file
was not cancelled at all but the user interface reflected a different
situation.

Now the controls of the UI element for the job will be changed ONLY when
the cancellation has happened and the web worker has properly notified
the fact to the Presenter.