Rewrite error notifications
Until now error notifications was a mix of in-place handling, a default
error handler, and direct calls to the user interface method for showing
errors to the end user. In addition to this, some errors were also shown
on console.

Now, ALL runtime errors are exceptions, and are thrown instead of just
being printed, so they can be handled, if at all needed in the future,
much easier. Once thrown, the errors are caught by the default exception
handler which, depending on the circumstances, will show an alert popup,
show the errors on the user interface, etc.

In addition to this, errors are now ALWAYS dumped to the console.

Because of this change, minor tweaks have been made to the HTML code.