Rewrite job handling and UI event handling
Events from the UI were handled using a subscriber pattern, which was a
limitation for job handling, excessively coupling the Presenter with the
View and with jobs, since all event handling for controls in the job UI
elements had to be made through the View itself.

Now, jobs are handled differently, have been decoupled by broadcasting
their events, and those events are handled by taking advantage of custom
events and bubbling.

In order to achieve this the jobs are now encapsulated separately.