Improve error notification UI
Until now the error notification was separated in two parts, one which
was used when the page was not entirely loaded and another, within the
UI class, for when the page was up and running.

While the error notification UI belongs to the View, it is not essential
to make it part of the UI class, in fact it does not make sense because
for the error notification to work the page has to be loaded BUT the UI
class does not need to be initialized or even created.