Move showVersion() out of UI class
This function works even if the UI class has not been created, and since
the version number retrieval and showing are now asynchronous, it is not
possible to know if the UI class has been already initialized unless a
global flag is implemented. It is much better to leave this function out
of the UI class.