Move showSlowModeStatus() out of UI class
This function works even if the UI class has not been created, and since
the slow mode status works asynchronously, it is not possible to know if
the UI class has been already initialized unless a global flag is added.
It is much better to leave this function out of the UI class.