Remove error showing fallback solution
Since the main JavaScript code is loaded using 'defer', it will not run
until the page is entirely loaded. So, the 'alert()' box that was being
used as fallback for showing errors before the page was fully loaded is
not needed.

In fact, errors happening BEFORE the main code is run will go unnoticed
anyway if they happen BEFORE the main error handling code is in place,
so the 'alert()' box was more or less useless.