Modify message handling in web worker
The switch/case that was being used was more than enough, an allowed for
minor code optimizations. Nonetheless, it was not very scalable and it
coupled the different handlers together since the message arguments were
processed before getting into the different case branches.

So, in order to make the handling more scalable and fully decouple the
different handlers, the previously used solution (separate handlers) is
used.