Modularize JavaScript code
The JavaScript code has been converted to modules. Right now there is no
real advantage, since there are no imports, but in the future there may
be imported scripts, strict mode is active implicitly, and all modules
are deferred by default, so in the end is a marginally better solution
overall, even if no imports are done now.

The only drawback is that it is not supported by old and marginally used
browsers, but quite a lot of the code would not work anyway with those
outdated or nearly unused browser, so it is not a deal breaker.