Improve custom events contract
Events being events, they are just strings, which is a weak contract but
it is the standard. The problem with this kind of contract is that typos
are quite difficult to find, and IDEs cannot really autocomplete or help
in any other way when writing related code.

Now, custom events are object properties and as such they are completed
by the IDE, warnings are issued for typos, etc. The contract is better,
and the only drawback is that event handling functions expect a string
and not an object property. This has been addressed, too, by assigning
string values for those properties.