Rename JobView class to Job
Even though this class is an abstraction of a job view, it is the only
class belonging to the View component having 'View' in its name, and it
is making the code a bit confusing to follow.

So, the class has been renamed and the rest of the code has been adapted
accordingly.