Refactor job state messages into Job class
Previously, the Presenter defined the state messages for the Job views,
including error messages, even preformatting them in HTML.

Now, all that rendering business logic has been moved into the Job class
which belonging to the View component is ultimately responsible for the
rendering of contents and deciding the proper messages to be shown.

This makes the code related to job state messages simpler and clearer.