Improve handling of slow mode
Previously, both the Presenter and the Web Worker stored the slow mode
state, and implemented part of the logic of switching it on and off.

Now, the main script and the Presenter do not store any such logic and
treat the indicator as a black box. It is the Web Worker responsibility
to handle the state of slow mode and report it so the indicator reflects
that state.