Remove showError() function
The only advantage of having that function was to make the defalt error
handler a bit shorter, but the disadvantage is that syntax errors which
happen before the default error handler is set up are missed.

Removing the function does not ensure a better behavior in that case but
minimizes the chances of missing early syntax errors.