Refactor some string literals into strings.js
Most if not all string literals are going to be refactored into a module
so the application is easier to localize and modify in the future.

There are, for now, three kind of strings: error messages, messages and
arbitrary strings.

Some string literals will be refactored as constants, in another module.