Decouple JavaScript from HTML and CSS
So far, JavaScript code was aware of the general structure of the HTML,
and about classes and ids used in CSS. Now, data attributes are used as
the contract between HTML and JavaScript, so styling selectors are not
used anymore in the JavaScript code.

Also, the selectors have been refactored into constants, so if the HTML
changes, to certain extent the JavaScript code will not, just the new
selectors will be needed for the code to work as-is.

Some HTML and CSS modifications have been necessary to sanitize all the
selectors and to make the decoupling possible.