Remove file reading cancel/retry logic
At first, being able to cancel and retry a file reading job looked like
a good idea, for the use case of being able to temporarily stop reading
a file for some reason. But the main reason why that should be needed is
that the file was added by mistake (it was too big, it does not have to
be processed right now, etc.), and in that case wanting to read the file
again at a later time does not seem to be a common scenario.

If a file was added by mistake, or it should be actually processed at a
later time, the job can be dismissed and the file added back again later
using the available controls, no need for a dedicated cancel and retry
logic (or a pause and resume one, for that matter).

Since having this logic increases code complexity and makes it harder to
debug properly, giving no real benefit back, it has been removed.