Check for missing jobs in the web worker
This is done to avoid nasty unhandled errors because of race conditions.

If for some reason (e.g. artificial delays introduced to test the UI) a
command on a job results on that job being deleted from the job registry
while there are some other pending commands, this can result on an error
because that job will not be defined when the pending command runs.