Remove 'slow mode'
This mode was needed for properly testing the UI. Since reading a file
can be VERY fast, it was very difficult to test the cancel/retry logic,
amongst other things. The main problem was that this slow mode consisted
of, basically, a busy wait loop. Not a great method.

Since that logic no longer exists, the only other UI element which would
benefit from a so-called 'slow mode' would be the job delete button. But
this button can be tested after a file has already been read, with only
an exception, and it's to simulate race conditions deleting a job whose
file is being read. This can be tested in many different ways, and using
a 'slow mode' using a busy wait loop is probably the worse.